requirements.txt
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
filelock==3.16.0
fonttools==4.53.1
fsspec==2024.9.0
idna==3.10
Jinja2==3.1.4
jiwer==3.0.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
noisereduce==3.0.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20231117
packaging==24.1
pillow==10.4.0
platformdirs==4.3.3
pooch==1.8.2
pycparser==2.22
pydub==0.25.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
rapidfuzz==3.9.7
regex==2024.9.11
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.13.2
threadpoolctl==3.5.0
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
